+++ project/model/position.py
@@ -1,14 +1,14 @@
 class Position:
-    def __init__(self, co_x: float, co_y: float):
-        self.__co_x: float = co_x
-        self.__co_y: float = co_y
+    def __init__(self, co_x: int, co_y: int):
+        self.__co_x: int = co_x
+        self.__co_y: int = co_y
 
     @property
-    def co_x(self) -> float:
+    def co_x(self) -> int:
         return self.__co_x
 
     @property
-    def co_y(self) -> float:
+    def co_y(self) -> int:
         return self.__co_y
 
     def __str__(self) -> str:
